Added obstacle collisions to the rays.

+++ Ray-Tracing-in-CPP/Circle.cpp
@@ -9,11 +9,11 @@ Circle::Circle(int x, int y, int r)
 
 void Circle::fillCircle(SDL_Renderer* renderer, Circle circle)
 {
-    double r2 = circle.r * circle.r;
+    float r2 = circle.r * circle.r;
 
-    for (double x = circle.x - circle.r; x <= circle.x + circle.r; x++)
+    for (float x = circle.x - circle.r; x <= circle.x + circle.r; x++)
     {
-        for (double y = circle.y - circle.r; y < circle.y + circle.r; y++)
+        for (float y = circle.y - circle.r; y < circle.y + circle.r; y++)
         {
             if ((x - circle.x) * (x - circle.x) + (y - circle.y) * (y - circle.y) <= r2)
             {
